Button connected example (on going)

+++ src/main/scala/agent/simulation/graphical/netlogo/component/NetlogoAgentComponent.scala
@@ -2,5 +2,5 @@ package agent.simulation.graphical.netlogo.component
 import scala.concurrent.duration._
 import agent.simulation.graphical.netlogo._
 
- abstract class NetlogoAgentComponent(netlogoModel : NetlogoModel)(maxTicks:Int = 1000)(fps: Int = 30) extends NetlogoAgent(netlogoModel)(maxTicks)(fps) {
+ abstract class NetlogoAgentComponent(netlogoModel : NetlogoModel)(maxTicks:Int = 1000)(fps: Float = 30f) extends NetlogoAgent(netlogoModel)(maxTicks)(fps) {
 }
